writing on computer
I wrote in Notepad on my computer, and it uploaded to GitHub on my
computer. The GitHub on my computer can be synced to the online/cloud
GitHub. So I can work through the GitHub file on my computer using
Notepad to save/commit/change files interchangeably. I believe from the
file through my computer and the online GitHub.